dev_spk

+++ SbsdbServer/Model/Entities/UserSettings.cs
@@ -6,9 +6,8 @@ namespace hb.SbsdbServer.Model.Entities
 {
     public partial class UserSettings
     {
-        public decimal Id { get; set; }
+        public long Id { get; set; }
         public string Userid { get; set; }
-        // Objekt wird als JSON-String abgelegt
         public UserSession Settings { get; set; }
-    }
+  }
 }
